Add vCard namespace.
I'm not really good at reading DTD so xmpp_*.h.in lists may not be
correct. For example, Psi uses attributes 'version' and 'prodid' but
as I understand the DTD, those should be elements. For the record, I
used the DTD included in XEP-0054. The bottom line is: all this needs a
review!

SVN Revision: 454