Use extended marker and driver versioning without SMP support.
Extended marker indicates that driver supports additional flags
like versioning and port level locking. Currently it was only
used when SMP support was enabled. Versioning seems like a nice
feature and has nothing to do with SMP (and is present in R12B5)
so enable it regardless of SMP support.